Refactored scaffold package
* Moved all New* functions to separated files
* Moved all New* test cases to associated test file
