+++ internal/scaffold/scaffold.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
-	"regexp"
 	"strings"
 	"time"
 
@@ -135,351 +134,5 @@ func RenderTemplates(renderer Renderer, templateFiles []TemplateFile, data any, 
 			return fmt.Errorf("failed to write file %s: %w", t.Destination, err)
 		}
 	}
-	return nil
-}
-
-func NewModule(opts Options) error {
-	if err := validateComponentName(opts.Name); err != nil {
-		return fmt.Errorf("invalid module name: %w", err)
-	}
-
-	baseDir := opts.TargetDir
-	if baseDir == "" {
-		cwd, err := os.Getwd()
-		if err != nil {
-			return fmt.Errorf("failed to get working directory: %w", err)
-		}
-		baseDir = cwd
-	}
-
-	moduleDir := filepath.Join(baseDir, opts.Name)
-
-	if _, err := os.Stat(moduleDir); err == nil {
-		if !opts.Force {
-			return fmt.Errorf("directory %s already exists, use --force to replace it", moduleDir)
-		}
-		if err := BackupDir(moduleDir); err != nil {
-			return fmt.Errorf("failed to back up existing directory: %w", err)
-		}
-		fmt.Printf("Renamed existing directory %s\n", moduleDir)
-	}
-
-	if err := os.MkdirAll(moduleDir, 0755); err != nil {
-		return fmt.Errorf("failed to create directory %s: %w", moduleDir, err)
-	}
-
-	meta := module.NewMetadata(opts.Name, opts.ForgeUser, opts.Author)
-	meta.License = opts.License
-	meta.Summary = opts.Summary
-	meta.Source = opts.Source
-
-	if err := meta.Write(filepath.Join(moduleDir, "metadata.json")); err != nil {
-		return fmt.Errorf("failed to write metadata.json: %w", err)
-	}
-
-	renderer := newRenderer(opts.TemplateDir)
-
-	templates := []TemplateFile{
-		{FileName: "module/manifests/init.pp", Destination: filepath.Join(moduleDir, "manifests", "init.pp")},
-		{FileName: "module/README.md", Destination: filepath.Join(moduleDir, "README.md")},
-		{FileName: "module/CHANGELOG.md", Destination: filepath.Join(moduleDir, "CHANGELOG.md")},
-		{FileName: "module/spec/class_spec.rb", Destination: filepath.Join(moduleDir, "spec", "classes", "init_spec.rb")},
-		{FileName: "module/Gemfile", Destination: filepath.Join(moduleDir, "Gemfile")},
-		{FileName: "module/Rakefile", Destination: filepath.Join(moduleDir, "Rakefile")},
-		{FileName: "module/gitignore", Destination: filepath.Join(moduleDir, ".gitignore")},
-		{FileName: "module/pdkignore", Destination: filepath.Join(moduleDir, ".pdkignore")},
-		{FileName: "module/rubocop.yml", Destination: filepath.Join(moduleDir, ".rubocop.yml")},
-		{FileName: "module/hiera.yaml", Destination: filepath.Join(moduleDir, "hiera.yaml")},
-		{FileName: "module/spec/spec_helper.rb", Destination: filepath.Join(moduleDir, "spec", "spec_helper.rb")},
-		{FileName: "module/spec/default_facts.yml", Destination: filepath.Join(moduleDir, "spec", "default_facts.yml")},
-		{FileName: "module/data/common.yaml", Destination: filepath.Join(moduleDir, "data", "common.yaml")},
-		{FileName: "common/gitkeep", Destination: filepath.Join(moduleDir, "data", ".gitkeep")},
-		{FileName: "common/gitkeep", Destination: filepath.Join(moduleDir, "examples", ".gitkeep")},
-		{FileName: "common/gitkeep", Destination: filepath.Join(moduleDir, "files", ".gitkeep")},
-		{FileName: "common/gitkeep", Destination: filepath.Join(moduleDir, "tasks", ".gitkeep")},
-		{FileName: "common/gitkeep", Destination: filepath.Join(moduleDir, "templates", ".gitkeep")},
-	}
-
-	data := struct {
-		ModuleName string
-		Author     string
-		License    string
-		ClassName  string
-	}{
-		ModuleName: opts.Name,
-		Author:     opts.Author,
-		License:    opts.License,
-		ClassName:  opts.Name,
-	}
-
-	if err := RenderTemplates(renderer, templates, data, opts.Force); err != nil {
-		return fmt.Errorf("failed to render templates: %w", err)
-	}
-
-	fmt.Printf("Created new module %s in %s\n", opts.Name, moduleDir)
-	return nil
-}
-
-func NewClass(opts ComponentOptions) error {
-	metadata, err := GetMetadata(opts.WorkDir)
-	if err != nil {
-		return fmt.Errorf("failed to get metadata: %w", err)
-	}
-
-	moduleName := metadata.ModuleName()
-
-	classFile, err := ConstructDestinationFilename(
-		opts.Name,
-		moduleName,
-		filepath.Join(opts.WorkDir, "manifests"),
-		".pp",
-	)
-	if err != nil {
-		return fmt.Errorf("failed to construct class file path: %w", err)
-	}
-
-	specFile, err := ConstructDestinationFilename(
-		opts.Name,
-		moduleName,
-		filepath.Join(opts.WorkDir, "spec", "classes"),
-		"_spec.rb",
-	)
-	if err != nil {
-		return fmt.Errorf("failed to construct spec file path: %w", err)
-	}
-
-	className := fmt.Sprintf("%s::%s", moduleName, opts.Name)
-
-	if _, err := os.Stat(classFile); err == nil {
-		return fmt.Errorf("class %s already exists: %s", className, classFile)
-	}
-
-	renderer := newRenderer(opts.TemplateDir)
-
-	templates := []TemplateFile{
-		{FileName: "class/class.pp", Destination: classFile},
-		{FileName: "class/class_spec.rb", Destination: specFile},
-	}
-
-	data := struct{ Name string }{Name: className}
-
-	fmt.Printf("creating class %s...\n", className)
-
-	if err := RenderTemplates(renderer, templates, data, false); err != nil {
-		return fmt.Errorf("failed to render templates: %w", err)
-	}
-
-	return nil
-}
-
-func NewDefinedType(opts ComponentOptions) error {
-	metadata, err := GetMetadata(opts.WorkDir)
-	if err != nil {
-		return fmt.Errorf("failed to get metadata: %w", err)
-	}
-
-	moduleName := metadata.ModuleName()
-
-	typeFile, err := ConstructDestinationFilename(
-		opts.Name,
-		moduleName,
-		filepath.Join(opts.WorkDir, "manifests"),
-		".pp",
-	)
-	if err != nil {
-		return fmt.Errorf("failed to construct defined_type file path: %w", err)
-	}
-
-	specFile, err := ConstructDestinationFilename(
-		opts.Name,
-		moduleName,
-		filepath.Join(opts.WorkDir, "spec", "defines"),
-		"_spec.rb",
-	)
-	if err != nil {
-		return fmt.Errorf("failed to construct defined_type test file path: %w", err)
-	}
-
-	typeName := fmt.Sprintf("%s::%s", moduleName, opts.Name)
-
-	if _, err := os.Stat(typeFile); err == nil {
-		return fmt.Errorf("defined_type %s already exists: %s", typeName, typeFile)
-	}
-	if _, err := os.Stat(specFile); err == nil {
-		return fmt.Errorf("defined_type %s test already exists: %s", typeName, specFile)
-	}
-
-	renderer := newRenderer(opts.TemplateDir)
-
-	templates := []TemplateFile{
-		{FileName: "type/defined_type.pp", Destination: typeFile},
-		{FileName: "type/defined_type_spec.rb", Destination: specFile},
-	}
-
-	data := struct{ Name string }{Name: typeName}
-
-	fmt.Printf("creating defined_type %s...\n", typeName)
-
-	if err := RenderTemplates(renderer, templates, data, false); err != nil {
-		return fmt.Errorf("failed to render templates: %w", err)
-	}
-
-	return nil
-}
-
-func NewFact(opts ComponentOptions) error {
-	// We only run this to see that we're in a valid module directory.
-	// Nothing is ever done with the metadata.
-	_, err := GetMetadata(opts.WorkDir)
-	if err != nil {
-		return fmt.Errorf("failed to get metadata: %w", err)
-	}
-
-	// Check to ensure opts.Name doesn't contain '::' as that would be invalid
-	// for a fact name.
-	if strings.Contains(opts.Name, "::") {
-		return fmt.Errorf("fact name cannot contain '::'")
-	}
-
-	// Check to see if a fact by given name already exists in the module, and
-	// return an error if it does.
-	factFileName := filepath.Join(opts.WorkDir, "lib", "facter", opts.Name+".rb")
-	if _, err := os.Stat(factFileName); err == nil {
-		return fmt.Errorf("fact %s already exists: %s", opts.Name, factFileName)
-	}
-
-	factTestFileName := filepath.Join(opts.WorkDir, "spec", "unit", "facter", opts.Name+"_spec.rb")
-	if _, err := os.Stat(factTestFileName); err == nil {
-		return fmt.Errorf("fact %s test already exists: %s", opts.Name, factTestFileName)
-	}
-
-	// Set up the templates and render them to disk at the appropriate locations.
-	fmt.Printf("creating fact %s...\n", opts.Name)
-	renderer := newRenderer(opts.TemplateDir)
-
-	templates := []TemplateFile{
-		{FileName: "fact/fact.rb", Destination: factFileName},
-		{FileName: "fact/fact_spec.rb", Destination: factTestFileName},
-	}
-
-	data := struct{ Name string }{Name: opts.Name}
-
-	if err := RenderTemplates(renderer, templates, data, false); err != nil {
-		return fmt.Errorf("failed to render templates: %w", err)
-	}
-
-	return nil
-}
-
-func NewFunction(opts ComponentOptions) error {
-	// Get metadata, we need the module name to construct the function name
-	metadata, err := GetMetadata(opts.WorkDir)
-	if err != nil {
-		return fmt.Errorf("failed to get metadata: %w", err)
-	}
-
-	moduleName := metadata.ModuleName()
-	functionName := fmt.Sprintf("%s::%s", moduleName, opts.Name)
-
-	functionFile, err := ConstructDestinationFilename(
-		opts.Name,
-		moduleName,
-		filepath.Join(opts.WorkDir, "functions"),
-		".pp",
-	)
-	if err != nil {
-		return fmt.Errorf("failed to construct function file path: %w", err)
-	}
-	if _, err := os.Stat(functionFile); err == nil {
-		return fmt.Errorf("function %s already exists: %s", functionName, functionFile)
-	}
-
-	specFile, err := ConstructDestinationFilename(
-		opts.Name,
-		moduleName,
-		filepath.Join(opts.WorkDir, "spec", "functions"),
-		"_spec.rb",
-	)
-	if err != nil {
-		return fmt.Errorf("failed to construct function test file path: %w", err)
-	}
-	if _, err := os.Stat(specFile); err == nil {
-		return fmt.Errorf("function %s test already exists: %s", functionName, specFile)
-	}
-
-	renderer := newRenderer(opts.TemplateDir)
-
-	templates := []TemplateFile{
-		{FileName: "function/function.pp", Destination: functionFile},
-		{FileName: "function/function_spec.rb", Destination: specFile},
-	}
-
-	data := struct{ Name string }{Name: functionName}
-
-	fmt.Printf("creating function %s...\n", functionName)
-
-	if err := RenderTemplates(renderer, templates, data, false); err != nil {
-		return fmt.Errorf("failed to render templates: %w", err)
-	}
-
-	return nil
-}
-
-func NewTask(opts ComponentOptions) error {
-	// Get the metadata, this is only done here to ensure we're in a valid
-	// module directory.'
-	_, err := GetMetadata(opts.WorkDir)
-	if err != nil {
-		return fmt.Errorf("failed to get metadata: %w", err)
-	}
-
-	// Validate that the task name meets the task name requirements
-	// [a-z][a-z0-9_]*
-	taskNamePattern := regexp.MustCompile(`^[a-z][a-z0-9_]*$`)
-	if !taskNamePattern.MatchString(opts.Name) {
-		return fmt.Errorf(
-			"task name %q does not match the allowed pattern (%s)",
-			opts.Name,
-			taskNamePattern.String(),
-		)
-	}
-
-	// Check to see if a task by given name already exists in the module, and
-	// return an error if it does.
-	taskFileName := filepath.Join(opts.WorkDir, "tasks", opts.Name+".sh")
-	if _, err := os.Stat(taskFileName); err == nil {
-		return fmt.Errorf(
-			"task %s already exists: %s",
-			opts.Name,
-			taskFileName,
-		)
-	}
-
-	// Check to see if a task metadata file exists in the module, and return an
-	// error if it does.
-	taskMetadataName := filepath.Join(opts.WorkDir, "tasks", opts.Name+".json")
-	if _, err := os.Stat(taskMetadataName); err == nil {
-		return fmt.Errorf(
-			"metadata file for task %s already exists: %s",
-			opts.Name,
-			taskMetadataName,
-		)
-	}
-
-	renderer := newRenderer(opts.TemplateDir)
-
-	templates := []TemplateFile{
-		{FileName: "task/task.sh", Destination: taskFileName},
-		{FileName: "task/metadata.json", Destination: taskMetadataName},
-	}
-
-	data := struct{}{}
-
-	fmt.Printf("creating task %s...\n", opts.Name)
-
-	if err := RenderTemplates(renderer, templates, data, false); err != nil {
-		return fmt.Errorf("failed to render templates: %w", err)
-	}
-
 	return nil
 }
